almost there
very weird stuff happening with setEnd and RegExpToFSA. Depending on how you run the test, it may pass or fail. No idea what's going on with that. There's too many swapping of pointers and items in sets and stuff, and it's a mess.

Maybe you should ditch the whole setEnd system. There must be a better way to do it.

Maybe Have no EndState and make a FSA class which contains the start and end state of a single regex. The end will be an empty state and you'll be able to just add next states to it. think it out idk